Set CLANG_WARN_SUSPICIOUS_IMPLICIT_CONVERSION to YES and fix errors

+++ GCJumpBar/GCJumpBarLabel.m
@@ -122,7 +122,7 @@ const NSInteger GCJumpBarLabelAccessoryMenuLabelTag = -1;
     else baseLeft = GCJumpBarLabelMargin; 
     
     if (self.image != nil) {
-        [self.image drawAtPoint:NSMakePoint(baseLeft, floorf(self.frame.size.height / 2 - self.image.size.height / 2)) 
+        [self.image drawAtPoint:NSMakePoint(baseLeft, floor(self.frame.size.height / 2 - self.image.size.height / 2)) 
                        fromRect:NSZeroRect 
                       operation:NSCompositeSourceOver 
                        fraction:1.0];
